+++ src/Rendering/Text.cpp
@@ -61,7 +61,7 @@ void Text::initialize(const std::string& fontFileName, double fontSize)
 	ftgl::mat4_set_identity(&mvp);
 }
 
-void Text::setWindowSize(int width, int height)
+void Text::setWindowSize(size_t width, size_t height)
 {
 	ftgl::mat4_set_orthographic(&mvp, 0, float(width), 0, float(height), -1, 1);
 }
