+++ src/Runners/ConsoleRunner.h
@@ -25,7 +25,7 @@ namespace Raycer
 
 	private:
 
-		void printProgress(const std::chrono::time_point<std::chrono::high_resolution_clock>& startTime, int totalPixelCount, int pixelsProcessed);
+		void printProgress(const std::chrono::time_point<std::chrono::high_resolution_clock>& startTime, size_t totalPixelCount, size_t pixelsProcessed);
 		void printProgressOpenCL(const std::chrono::time_point<std::chrono::high_resolution_clock>& startTime);
 		std::string humanizeNumberDecimal(double value);
 		std::string humanizeNumberBytes(double value);
