+++ src/Runners/WindowRunnerStates/DefaultState.cpp
@@ -155,20 +155,20 @@ void DefaultState::render(double timeStep, double interpolation)
 
 	if (settings.window.showInfoText)
 	{
-		text.drawText(5.0, runner.getWindowHeight() - 3 * settings.window.defaultFontSize, Color(255, 255, 255, 255), tfm::format("Pos: (%.2f, %.2f, %.2f)", scene.camera.position.x, scene.camera.position.y, scene.camera.position.z));
-		text.drawText(5.0, runner.getWindowHeight() - 4 * settings.window.defaultFontSize - 2, Color(255, 255, 255, 255), tfm::format("Rot: (%.2f, %.2f, %.2f)", scene.camera.orientation.pitch, scene.camera.orientation.yaw, scene.camera.orientation.roll));
-		text.drawText(5.0, runner.getWindowHeight() - 5 * settings.window.defaultFontSize - 4, Color(255, 255, 255, 255), tfm::format("Pix: (%d, %d)", runner.getMouseInfo().framebufferX, runner.getMouseInfo().framebufferY));
-		text.drawText(5.0, runner.getWindowHeight() - 6 * settings.window.defaultFontSize - 6, Color(255, 255, 255, 255), tfm::format("Mov: %s", scene.camera.hasMoved()));
+		text.drawText(5.0, double(runner.getWindowHeight() - 3 * settings.window.defaultFontSize), Color(255, 255, 255, 255), tfm::format("Pos: (%.2f, %.2f, %.2f)", scene.camera.position.x, scene.camera.position.y, scene.camera.position.z));
+		text.drawText(5.0, double(runner.getWindowHeight() - 4 * settings.window.defaultFontSize - 2), Color(255, 255, 255, 255), tfm::format("Rot: (%.2f, %.2f, %.2f)", scene.camera.orientation.pitch, scene.camera.orientation.yaw, scene.camera.orientation.roll));
+		text.drawText(5.0, double(runner.getWindowHeight() - 5 * settings.window.defaultFontSize - 4), Color(255, 255, 255, 255), tfm::format("Pix: (%d, %d)", runner.getMouseInfo().framebufferX, runner.getMouseInfo().framebufferY));
+		text.drawText(5.0, double(runner.getWindowHeight() - 6 * settings.window.defaultFontSize - 6), Color(255, 255, 255, 255), tfm::format("Mov: %s", scene.camera.hasMoved()));
 	}
 }
 
-void DefaultState::windowResized(int width, int height)
+void DefaultState::windowResized(size_t width, size_t height)
 {
 	(void)width;
 	(void)height;
 }
 
-void DefaultState::framebufferResized(int width, int height)
+void DefaultState::framebufferResized(size_t width, size_t height)
 {
 	scene.camera.setImagePlaneSize(width, height);
 }
